Splitting storage, access, and orbis services and removing vector search
